requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=1.0.1
BeautifulSoup
lxml
pandas
numpy
hyperlink
requests-html
selenium